requirements.txt
numpy
scipy
setuptools
pillow